+++ Classes/EventsController.h
@@ -4,4 +4,5 @@
 - (id)initWithEvents:(GHEvents *)events;
 - (void)openEventItem:(id)eventItem;
 - (void)refreshLastUpdate;
+- (void)refreshIfRequired;
 @end