+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes:
 
 * Check the users assignment status for editing issues and pulls
 * Separate section for forks in personal repositories list
+* Better handling of events reloading
 
 Bugfixes: